[REST] Housed calls updated
-Updated the calls for Housed
-New method in GroupLeader Service

+++ src/resources/HousedResource.java
@@ -1,12 +1,8 @@
 package resources;
 
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
-import javax.annotation.PostConstruct;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
@@ -15,33 +11,21 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 
-import model.Booking;
-import model.Guest;
+import model.GroupLeader;
 import model.Housed;
-import model.questura.Country;
-import model.questura.HousedType;
-import model.questura.Municipality;
 
-import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.SolrServer;
-import org.apache.solr.client.solrj.SolrServerException;
-import org.apache.solr.client.solrj.response.QueryResponse;
-import org.apache.solr.client.solrj.response.FacetField.Count;
-import org.apache.solr.common.SolrDocument;
-import org.apache.solr.common.SolrDocumentList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import com.sun.jersey.api.NotFoundException;
-
 import service.BookingService;
-import service.GuestService;
+import service.GroupLeaderService;
 import service.HousedService;
 import service.StructureService;
+
+import com.sun.jersey.api.NotFoundException;
 
 @Path("/housed/")
 @Component
@@ -54,131 +38,38 @@ public class HousedResource {
     @Autowired
     private BookingService bookingService = null;
     @Autowired
-    private SolrServer solrServerGuest = null;
-   
+    private GroupLeaderService groupLeaderService = null;
     
-//    @PostConstruct
-//    public void init(){
-//    	List<Guest> guests = null;
-//    	
-//    	guests = this.getGuestService().findAll();
-//    	try {
-//			this.getSolrServerGuest().addBeans(guests);
-//			this.getSolrServerGuest().commit();
-//		} catch (SolrServerException e) {			
-//			e.printStackTrace();
-//		} catch (IOException e) {
-//			e.printStackTrace();
-//		}
-//    }
     @GET
-    @Path("booking/{idBooking}/search/{start}/{rows}")
+    @Path("booking/{idBooking}")
     @Produces({MediaType.APPLICATION_JSON})
-    public List<Housed> search(@PathParam("idBooking") Integer idBooking, @PathParam("start") Integer start, @PathParam("rows") Integer rows) {
+    public List<Housed> findHousedByIdBooking(@PathParam("idBooking") Integer idBooking) {
     	List<Housed> ret = null;
     	
     	ret = this.getHousedService().findHousedByIdBooking(idBooking);
     	return ret;
     }
-    
-    
-//    @GET
-//    @Path("structure/{idStructure}/search/{start}/{rows}")
-//    @Produces({MediaType.APPLICATION_JSON})
-//    public List<Guest> search(@PathParam("idStructure") Integer idStructure,@PathParam("start") Integer start,@PathParam("rows") Integer rows, @QueryParam("term") String term){
-//        List<Guest> guests = null;
-//        SolrQuery query = null;
-//        QueryResponse rsp = null;
-//        SolrDocumentList solrDocumentList = null;
-//        SolrDocument solrDocument = null;
-//        Guest aGuest = null;
-//        Integer id;             
-//       
-//        if(term.trim().equals("")){
-//        	term = "*:*";
-//        }
-//        term = term + " AND id_structure:" + idStructure.toString();
-//        query = new SolrQuery();   		
-//        query.setQuery(term);
-//        query.setStart(start);
-//        query.setRows(rows);
-//              
-//        try {
-//			rsp = this.getSolrServerGuest().query(query);
-//			
-//		} catch (SolrServerException e) {
-//			e.printStackTrace();			
-//		}
-//
-//        guests = new ArrayList<Guest>();
-//        if(rsp!=null){
-//    	   solrDocumentList = rsp.getResults();
-//           for(int i = 0; i <solrDocumentList.size(); i++){
-//        	   solrDocument = solrDocumentList.get(i);
-//        	   id = (Integer)solrDocument.getFieldValue("id");
-//        	  // System.out.println("----> "+solrDocument.getFieldValues("text")+" <-----");
-//        	   aGuest = this.getGuestService().findGuestById(id);
-//        	   guests.add(aGuest);
-//           }  
-//       }       
-//       return guests;          
-//    }
-    
-//    @GET
-//   	@Path("structure/{idStructure}/suggest")
-//   	@Produces({ MediaType.APPLICATION_JSON })
-//   	public List<String> suggest(@PathParam("idStructure") Integer idStructure,
-//   			@QueryParam("term") String term) {
-//   		SolrQuery query = null;
-//   		QueryResponse rsp = null;
-//   		List<String> ret = null;
-//   		List<Count> values = null;
-//   		
-//   		query = new SolrQuery();
-//   		query.setFacet(true);
-//   		query.setQuery("*:* AND id_structure:" + idStructure.toString());
-//
-//   		query.addFacetField("text");
-//   		term = term.toLowerCase();
-//   		query.setFacetPrefix(term);
-//
-//   		try {
-//   			rsp = this.getSolrServerGuest().query(query);
-//   		} catch (SolrServerException e) {
-//   			e.printStackTrace();
-//   		}
-//   		ret = new ArrayList<String>();
-//
-//   		if (rsp != null) {
-//   			values = rsp.getFacetField("text").getValues();
-//   			if(values!=null){
-//   				for(Count each: values){
-//   					if(each.getCount()>0){
-//   						ret.add(each.getName());
-//   					}
-//   				}	
-//   			}					
-//   		}
-//   		return ret;
-//   	}
-//    
-    @GET
-    @Path("{id}")
-    @Produces({MediaType.APPLICATION_JSON})
-    public Housed getHoused(@PathParam("id") Integer id){
-    	Housed aHoused = null;
-       
-    	aHoused = this.getHousedService().findHousedById(id);
-        return aHoused;   
-    }
  
     @POST	
 	@Consumes({MediaType.APPLICATION_JSON})
 	@Produces({MediaType.APPLICATION_JSON}) 
-	public Housed insertHoused(Housed housed){
+	public Integer insertHoused(Map map){
+    	Housed housed = null;
+    	
+    	Integer id_booking = null;
+		Integer id_guest = null;
+ 		Integer id = null;
+ 		
+ 		id_booking = (Integer)map.get("id_booking");
+		id_guest = (Integer)map.get("id_guest");
+
+		housed = new Housed();
+		housed.setId_booking(id_booking);
+		housed.setId_guest(id_guest);
  		
  		this.getHousedService().insertHoused(housed);
- 		return housed;
+ 		id = housed.getId();
+ 		return id;
 	}
     
     
@@ -186,31 +77,48 @@ public class HousedResource {
     @Path("{id}")
     @Consumes({MediaType.APPLICATION_JSON})
     @Produces({MediaType.APPLICATION_JSON})
-    public Housed update(Housed housed) {  
+    public Integer update(Map map) {
+    	Integer id;
+    	Integer id_booking = null;
+		Integer id_guest = null;
+		Housed housed = null;
+		
+		id = (Integer)map.get("id");
+		id_booking = (Integer)map.get("id_booking");
+		id_guest = (Integer)map.get("id_guest");
+		
+		housed = this.getHousedService().findHousedById(id);
+		//TODO: date format to be set
+		housed.setId_guest(id_guest);
     	
     	this.getHousedService().updateHoused(housed);
-        return housed;
+        return id;
     }
    
     @DELETE
     @Path("{id}")
     @Produces({MediaType.APPLICATION_JSON})   
     public Integer delete(@PathParam("id") Integer id){
-    	Integer count = 0;		
+    	Integer count = 0;
+    	Housed housed = null;
+    	Integer id_booking;
+    	GroupLeader groupLeader = null;
 		
 //		if(this.getBookingService().countGroupLeaderByIdHoused(id) > 0){
 //			throw new NotFoundException("The housed you are trying to delete is the leader of the group associated with this booking" +
 //					" Please change the group/family leader first.");
 //		}
+    	
+    	housed = this.getHousedService().findHousedById(id);
+    	id_booking = housed.getId_booking();
+    	groupLeader = this.getGroupLeaderService().findGroupLeaderByIdBookingAndIdHoused(id_booking, id);
+    	if (groupLeader != null) {
+    		throw new NotFoundException("The housed you are trying to delete is the leader of the group associated with this booking" +
+    										"Please change the group/family leader first.");
+    	}
 		count = this.getHousedService().deleteHoused(id);
-		if(count == 0){
-			throw new NotFoundException("Error: the housed has NOT been deleted");
-		}
 		return count;
-    }   
-   
-    
-    
+    }  
     
     public HousedService getHousedService() {
 		return housedService;
@@ -230,11 +138,11 @@ public class HousedResource {
 	public void setBookingService(BookingService bookingService) {
 		this.bookingService = bookingService;
 	}
-	public SolrServer getSolrServerGuest() {
-		return solrServerGuest;
+	public GroupLeaderService getGroupLeaderService() {
+		return groupLeaderService;
 	}
-	public void setSolrServerGuest(SolrServer solrServerGuest) {
-		this.solrServerGuest = solrServerGuest;
-	}  	
+	public void setGroupLeaderService(GroupLeaderService groupLeaderService) {
+		this.groupLeaderService = groupLeaderService;
+	}	
 
 }
